add knn classifier, standardise sklearn methods
