+++ openproblems/tasks/label_projection/methods/__init__.py
@@ -2,3 +2,4 @@ from .logistic_regression import logistic_regression_log_cpm
 from .logistic_regression import logistic_regression_scran
 from .mlp import mlp_log_cpm
 from .mlp import mlp_scran
+from .knn_classifier import knn_classifier_log_cpm, knn_classifier_scran
